requirements.txt
opencv-python
numpy
pillow
pytesseract
tensorflow
keras
pandas 